requirements.txt
django
markdown
django-filter
requests
django-crispy-forms
psycopg2-binary
